# app/swagger.yaml
swagger: '2.0'
info:
  title: "Your API"
  description: "This is an API"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

    
#------------------------------------------STILL NEED TO ADD PATHS AND DEFINITIONS FOR OTHER ROUTES EX MECHS, VEHICLES, ETC
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Customer login"
      description: "Authenticates a customer and returns an auth token."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/LoginPayload"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
        401:
          description: "Invalid credentials"
  /customers:
    post:
      tags:
        - Customers
      summary: "Create a customer"
      description: "Creates a new customer account."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/CustomerResponse"

    get:
      tags:
        - Customers
      summary: "Get all customers"
      description: "Returns paginated list of customers. Token required."
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          type: integer
          required: false
        - in: query
          name: per_page
          type: integer
          required: false
      responses:
        200:
          description: "Paginated customers"
          schema:
            $ref: "#/definitions/PaginatedCustomers"
  /customers/{id}:
    put:
      tags:
        - Customers
      summary: "Update customer"
      description: "Updates a customer by ID. Token required."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/CustomerResponse"
        404:
          description: "Customer not found"

    delete:
      tags:
        - Customers
      summary: "Delete customer"
      description: "Deletes a customer by ID. Token required."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: "Customer deleted"
        404:
          description: "Customer not found"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Get my service tickets"
      description: "Returns all service tickets for the logged-in customer."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of service tickets"
          schema:
            $ref: "#/definitions/CustomerTickets"


#mechanics 
...
definitions:
  LoginPayload:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  LoginResponse:
    type: object
    properties:
      status:
        type: string
      auth_token:
        type: string
  CreateCustomerPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
    required:
      - name
      - email
      - password
  UpdateCustomerPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
  PaginatedCustomers:
    type: object
    properties:
      customers:
        type: array
        items:
          $ref: "#/definitions/CustomerResponse"
      total:
        type: integer
      pages:
        type: integer
      current_page:
        type: integer
  CustomerTickets:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        service_date:
          type: string
        cost:
          type: number
        vehicle_id:
          type: integer

# mechanics
...